+++ NuclearSegmentation/yousef_core/clustering_2D/Local_Max_Clust_2D.cpp
@@ -22,7 +22,7 @@ float get_maximum(float** A, int r1, int r2, int c1, int c2, int* rx, int* cx)
     return mx;
 }
 
-int local_max_clust_2D(float* im_vals, int r, int c, double scale, int* out1, int* seeds_x, int* seeds_y, int num_seeds, int* bin_image)
+int local_max_clust_2D(float* im_vals, int r, int c, double scale, unsigned short* out1, int* seeds_x, int* seeds_y, int num_seeds, unsigned short* bin_image)
 {		
     float** im;    
     float** local_max_im;
