+++ NuclearSegmentation/yousef_core/clustering_3D/Local_Max_Clust_3D_CPP.cpp
@@ -44,7 +44,7 @@ void get_maximum(float* A, int r1, int r2, int c1, int c2, int z1, int z2, int* 
 }
 
 
-void local_max_clust_3D(float* im_vals, int* local_max_vals, int* bImg, int* out1, int r, int c, int z, int scale_xy, int scale_z)
+void local_max_clust_3D(float* im_vals, int* local_max_vals, unsigned short* bImg, unsigned short* out1, int r, int c, int z, int scale_xy, int scale_z)
 {  
 	//im_vals is the Laplacian of Gaussian
 	//local_max_vals is the seed points (local maximum) with foreground seeds assigned an id > 0 and background seeds id == -1
