+++ NuclearSegmentation/yousef_core/clustering_3D/local_max_clust_3D.h
@@ -19,6 +19,6 @@ limitations under the License.
 #include <iostream>
 #include <algorithm>
 
-void local_max_clust_3D(float* im_vals, int* local_max_vals, int* bImg, int* out1, int r, int c, int z, int scale_xy, int scale_z);
+void local_max_clust_3D(float* im_vals, int* local_max_vals, unsigned short* bImg, unsigned short* out1, int r, int c, int z, int scale_xy, int scale_z);
 
 #endif
