+++ NuclearSegmentation/yousef_core/seed_detection_3D/seedsdetection_3D.h
@@ -16,6 +16,6 @@ limitations under the License.
 #ifndef SEEDSDETECTION_3D_H
 #define SEEDSDETECTION_3D_H
 
-int Seeds_Detection_3D( float* IM, float* IM_out, int* IM_bin, int r, int c, int z, double sigma_min, double sigma_max, double scale_xy, double scale_z,int sampl_ratio,int* bImg, int UseDistMap);
+int Seeds_Detection_3D( float* IM, float* IM_out, int* IM_bin, int r, int c, int z, double sigma_min, double sigma_max, double scale_xy, double scale_z,int sampl_ratio,unsigned short* bImg, int UseDistMap);
 
 #endif
